requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
transformers
torch
sentence-transformers
pydantic
requests
langchain
llama-index
haystack
